refactor: reorganize `clean_pdfminer_inner_elements()`
